# docs/openapi.yml
openapi: "3.0.0"
info:
  title: The Fiagram API
  description: The REST Full API specification for the Fiagram project
  version: 1.0.0
servers:
  - url: http://api.fiagram.com/v1
    description: Production server

  - url: http://localhost:8080/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /profile/me:
    get:
      tags: [Profile]
      summary: Get current account information
      operationId: getProfileMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current account information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileMeResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    put:
      tags: [Profile]
      summary: Update current account information
      operationId: updateProfileMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileMeRequest"
      responses:
        "200":
          description: Account information updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileMeResponse"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profile/me/password:
    put:
      tags: [Profile]
      summary: Update current account password
      operationId: updateProfilePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordRequest"
      responses:
        "204":
          description: Password updated successfully
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profile/webhooks:
    get:
      tags: [Profile]
      summary: List the account's webhooks
      operationId: getProfileWebhooks
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: List of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    post:
      tags: [Profile]
      summary: Create a new webhook
      operationId: createProfileWebhook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Webhook"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /profile/webhooks/{webhookId}:
    get:
      tags: [Profile]
      summary: Get a specific webhook
      operationId: getProfileWebhook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "200":
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    put:
      tags: [Profile]
      summary: Update a webhook
      operationId: updateProfileWebhook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Webhook"
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    delete:
      tags: [Profile]
      summary: Delete a webhook
      operationId: deleteProfileWebhook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "204":
          description: Webhook deleted
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------- Strategy
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: Max items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      description: Items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    WebhookId:
      name: webhookId
      in: path
      required: true
      description: The unique ID of the webhook
      schema:
        type: integer
        format: uint64
        minimum: 1
...
  responses:
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          example: UNAUTHORIZED
        message:
          type: string
          example: Unauthorized
        details:
          type: object
          additionalProperties: true
    Password:
      type: string
      format: password
      writeOnly: true
      minLength: 8
      maxLength: 72
      description: |
        8-72 characters, including at least one uppercase,
        one lowercase, one digit, and one special character; no whitespace.
      pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9])\S{8,72}$'
      example: Str0ngPassw0rd!
    Username:
      type: string
      minLength: 5
      maxLength: 20
      description: |
        5–20 characters; letters and numbers required;
        optional single dots or underscores allowed between characters;
        cannot start or end with dot/underscore; no consecutive dots or underscores.
      pattern: "^(?=.{5,20}$)(?!.*[_.]{2})[a-zA-Z0-9]+([._]?[a-zA-Z0-9]+)*$"
      example: member1
    Fullname:
      type: string
      minLength: 5
      maxLength: 200
      example: Nguyễn Vũ Hoàng Anh
    Email:
      type: string
      pattern: '^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$'
      example: user1@example.com
    PhoneNumber:
      type: string
      description: the phone number of a created account
      pattern: '^\d{6,15}$'
      example: "0123456789"
    Role:
      type: string
      enum:
        - none
        - admin
        - member
      example: member
      default: member
    Account:
      type: object
      required: [username, email, role, fullname]
      properties:
        username:
          $ref: "#/components/schemas/Username"
        fullname:
          $ref: "#/components/schemas/Fullname"
        email:
          $ref: "#/components/schemas/Email"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
        role:
          $ref: "#/components/schemas/Role"
...
    ProfileMeResponse:
      type: object
      additionalProperties: false
      required: [account]
      properties:
        account:
          $ref: "#/components/schemas/Account"
    UpdateProfileMeRequest:
      type: object
      additionalProperties: false
      properties:
        fullname:
          $ref: "#/components/schemas/Fullname"
        email:
          $ref: "#/components/schemas/Email"
        phoneNumber:
          $ref: "#/components/schemas/PhoneNumber"
    UpdatePasswordRequest:
      type: object
      additionalProperties: false
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          $ref: "#/components/schemas/Password"
        newPassword:
          $ref: "#/components/schemas/Password"
    Webhook:
      type: object
      additionalProperties: false
      required: [name, url]
      properties:
        id:
          type: integer
          format: uint64
          minLength: 1
          maxLength: 26
          pattern: "^(?=.{5,20}$)(?!.*[_.]{2})[a-zA-Z0-9]+([._]?[a-zA-Z0-9]+)*$"
          readOnly: true
          description: Server-generated unique ID
          example: 42
        name:
          type: string
        url:
          type: string

    # -------------------------------- Strategy
...